refactor(hub): extract auth, setup, JWT, password/PIN handlers into auth.go
Pure code move — no behavior change. main.go was 4344 lines; pulling
the auth-related block into its own file leaves 3768 in main.go and
593 in auth.go.

Moved (still package main, same globals — db, jwtSecret, rateLimiter):
  - ensureDefaultAdmin
  - setupTokenValue var, generateSetupToken
  - handleSetupStatus, handleSetup
  - handleLogin
  - jwtMiddleware
  - rotationAllowlist var, credentialRotationMiddleware
  - extractUserIDFromRequest
  - loadOrGenerateJWTSecret, generateRandomSecret, warnDefaultPassword
  - handleChangePassword
  - verifyTerminalPIN, handleChangePIN
  - handleSSEToken

Left in main.go: tokenRedactingWriter (logging), generateFirstRunToken
(install-token bootstrap — moves with agent enrollment refactor),
generateTerminalBrowserToken (terminal-session PIN — moves with terminal
extraction).

go test -count=1 ./...  green 3 runs in a row.
